Move uninterruptible atomics to guest staging
Extract the seven atomic helpers from the mixed core utility into a guest-owned staging class while preserving their implementations.

Add the module access required by the relocated helpers.
