+++ substratevm/mx.substratevm/suite.py
@@ -1468,6 +1468,9 @@ suite = {
                 "SVM_SHARED",
             ],
             "requiresConcealed" : {
+                "java.base" : [
+                    "jdk.internal.misc",
+                ],
                 "jdk.internal.vm.ci" : [
                     "jdk.vm.ci.meta",
                     "jdk.vm.ci.meta.annotation",
