+++ substratevm/src/com.oracle.svm.hosted.test/src/com/oracle/svm/hosted/test/VerifyReflectionUsage.java
@@ -156,11 +156,6 @@ public class VerifyReflectionUsage extends VerifyReflectionUsageBase {
                     clazz("com.oracle.svm.core.jdk.resources.MissingResourceRegistrationUtils"),
                     clazz("com.oracle.svm.core.jdk.SecurityProvidersSupport"),
                     clazz("com.oracle.svm.core.jdk.StackAccessControlContextVisitor"),
-                    clazz("com.oracle.svm.core.jdk.UninterruptibleUtils$AtomicBoolean"),
-                    clazz("com.oracle.svm.core.jdk.UninterruptibleUtils$AtomicInteger"),
-                    clazz("com.oracle.svm.core.jdk.UninterruptibleUtils$AtomicLong"),
-                    clazz("com.oracle.svm.core.jdk.UninterruptibleUtils$AtomicPointer"),
-                    clazz("com.oracle.svm.core.jdk.UninterruptibleUtils$AtomicReference"),
                     clazz("com.oracle.svm.core.jfr.events.ThreadParkEvent"),
                     clazz("com.oracle.svm.core.jfr.traceid.JfrTraceId"),
                     clazz("com.oracle.svm.core.jni.access.JNIAccessibleMember"),
